+++ app/containers/memberList.js
@@ -2,30 +2,42 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { currentChannelMembersSelector } from '../selectors'
 
-function MemberList ({ members = [] }) {
+import {
+  showProfilePanel
+} from '../actions'
+import Avatar from './avatar'
+
+const mapDispatchToProps = dispatch => ({
+  showProfilePanel: ({ addr, userKey }) => dispatch(showProfilePanel({ addr, userKey }))
+})
+
+function MemberList (props) {
+  function onClickUser (user) {
+    props.showProfilePanel({
+      addr: props.addr,
+      userKey: user.key
+    })
+  }
+
   return (
-    <div className='panel'>
-      <div className='panel__header'>
-        Channel Members
-      </div>
-      <div className='panel__content'>
-        {members.map((user) =>
-          <div key={user.key} className='collection__item'>
-            <div className='collection__item__icon'>
-              {!!user.online &&
-                <img alt='Online' src='static/images/icon-status-online.svg' />}
-              {!user.online &&
-                <img alt='Offline' src='static/images/icon-status-offline.svg' />}
-            </div>
+    <>
+      {props.members.map((user) =>
+        <div key={user.key} className='collection__item' onClick={() => onClickUser(user)} title={user.key}>
+          <div className='collection__item__icon'>
             {!!user.online &&
-              <div className='collection__item__content active'>{user.name || user.key.substring(0, 6)}</div>}
+              <img alt='Online' src='static/images/icon-status-online.svg' />}
             {!user.online &&
-              <div className='collection__item__content'>{user.name || user.key.substring(0, 6)}</div>}
-            <div className='collection__item__handle' />
+              <img alt='Offline' src='static/images/icon-status-offline.svg' />}
           </div>
-        )}
-      </div>
-    </div>
+          <Avatar name={user.key} scale={2} />
+          {!!user.online &&
+            <div className='collection__item__content active'>{user.name || user.key.substring(0, 6)}</div>}
+          {!user.online &&
+            <div className='collection__item__content'>{user.name || user.key.substring(0, 6)}</div>}
+          <div className='collection__item__handle' />
+        </div>
+      )}
+    </>
   )
 }
 
@@ -33,4 +45,4 @@ export default connect(state => {
   return {
     members: currentChannelMembersSelector(state)
   }
-}, null)(MemberList)
+}, mapDispatchToProps)(MemberList)