+++ app/containers/sidebar.js
@@ -40,14 +40,14 @@ const mapDispatchToProps = dispatch => ({
   setUsername: ({ addr, username }) => dispatch(setUsername({ addr, username })),
   showCabalSettings: ({ addr }) => dispatch(showCabalSettings({ addr })),
   showChannelBrowser: ({ addr }) => dispatch(showChannelBrowser({ addr })),
-  showProfilePanel: ({ addr, user }) => dispatch(showProfilePanel({ addr, user })),
+  showProfilePanel: ({ addr, userKey }) => dispatch(showProfilePanel({ addr, userKey })),
   viewChannel: ({ addr, channel }) => dispatch(viewChannel({ addr, channel }))
 })
 
 function UserMenu (props) {
   useEffect(() => {
-    console.warn('UserMenu', props)
-  }, [props.nick])
+    // console.warn('UserMenu', props)
+  }, [props.peer])
 
   return (
     <Menu id='user_menu' theme={theme.dark}>
@@ -123,17 +123,17 @@ class SidebarScreen extends React.Component {
   onClickUser (user) {
     this.props.showProfilePanel({
       addr: this.props.addr,
-      user
+      userKey: user.key
     })
   }
 
   onContextMenu (peer, e) {
-    e.preventDefault()
+    // e.preventDefault()
     // contextMenu.show({
     //   id: 'user_menu',
     //   event: e,
     //   props: {
-    //     nick: nick
+    //     peer
     //   }
     // })
   }
@@ -162,8 +162,14 @@ class SidebarScreen extends React.Component {
 
   sortUsers (users) {
     return users.sort((a, b) => {
+      if (a.isHidden() && !b.isHidden()) return 1
+      if (b.isHidden() && !a.isHidden()) return -1
       if (a.online && !b.online) return -1
       if (b.online && !a.online) return 1
+      if (a.isAdmin() && !b.isAdmin()) return -1
+      if (b.isAdmin() && !a.isAdmin()) return 1
+      if (a.isModerator() && !b.isModerator()) return -1
+      if (b.isModerator() && !a.isModerator()) return 1
       if (a.name && !b.name) return -1
       if (b.name && !a.name) return 1
       if (a.name && b.name) return a.name.toLowerCase() < b.name.toLowerCase() ? -1 : 1
@@ -188,21 +194,23 @@ class SidebarScreen extends React.Component {
   }
 
   render () {
-    var self = this
     const { addr, cabal, settings } = this.props
     const cabalLabel = settings.alias || addr
-
-    const favorites = (settings['favorite-channels'] || []).sort()
-    const channels = cabal.channelsJoined.slice().sort().filter(x => !favorites.includes(x))
+    const channelsJoined = cabal.channelsJoined?.slice().sort() || []
+    const favorites = channelsJoined.filter(channel => (settings['favorite-channels'] || []).includes(channel))
+    const channels = channelsJoined.filter(channel => !favorites.includes(channel))
     const users = this.sortUsers(Object.values(cabal.users) || [])
     const deduplicatedNicks = this.deduplicatedNicks(users)
     const onlineCount = users.filter(i => !!i.online).length
     const userkey = cabal.userkey
     const username = cabal.username
+    const unreadNonFavoriteMessageCount = Object.entries((this.props.channelMessagesUnread || {})).reduce((total, value) => {
+      return (value[1] && channels.includes(value[0])) ? (total + value[1]) : total
+    }, 0)
     return (
       <div className='client__sidebar'>
         <div className='sidebar'>
-          <div className='session' onClick={self.onClickCabalSettings.bind(self, cabal.addr)}>
+          <div className='session' onClick={this.onClickCabalSettings.bind(this, cabal.addr)}>
             <div className='session__avatar'>
               <div className='session__avatar__img'>
                 <Avatar name={userkey} />
@@ -210,11 +218,11 @@ class SidebarScreen extends React.Component {
             </div>
             <div className='session__meta'>
               <h1>{cabalLabel}</h1>
-              <h2 onClick={self.onClickUsername.bind(self)}>
+              <h2 onClick={this.onClickUsername.bind(this)}>
                 {username}
               </h2>
             </div>
-            <div className='session__configuration'>
+            <div className='session__configuration' title='Settings for this Cabal'>
               <img src='static/images/icon-sidebarmenu.svg' />
             </div>
           </div>
@@ -225,10 +233,15 @@ class SidebarScreen extends React.Component {
                   <div className='collection__heading__title__container'>
                     <span
                       className={`collection__toggle ${this.props.settings['sidebar-hide-favorites'] ? 'collection__toggle__off' : 'collection__toggle__on'}`}
-                      onClick={self.onToggleCollection.bind(self, 'favorites')}
+                      onClick={this.onToggleCollection.bind(this, 'favorites')}
                     >▼
                     </span>
-                    <div className='collection__heading__title'>Starred</div>
+                    <div
+                      className='collection__heading__title'
+                      onClick={this.onToggleCollection.bind(this, 'favorites')}
+                    >
+                      Starred
+                    </div>
                   </div>
                 </div>
                 {!this.props.settings['sidebar-hide-favorites'] && this.sortByProperty(favorites).map((channel) =>
@@ -246,17 +259,23 @@ class SidebarScreen extends React.Component {
                 <div className='collection__heading__title__container'>
                   <span
                     className={`collection__toggle ${this.props.settings['sidebar-hide-channels'] ? 'collection__toggle__off' : 'collection__toggle__on'}`}
-                    onClick={self.onToggleCollection.bind(self, 'channels')}
+                    onClick={this.onToggleCollection.bind(this, 'channels')}
                   >▼
                   </span>
                   <div
                     className='collection__heading__title collection__heading__title__channelBrowserButton'
-                    onClick={self.onClickChannelBrowser.bind(self, cabal.addr)}
-                    title='Browse and join all channels'
-                  >Channels
+                    onClick={this.onToggleCollection.bind(this, 'channels')}
+                  >
+                    Channels
+                    {this.props.settings['sidebar-hide-channels'] && unreadNonFavoriteMessageCount > 0 &&
+                      <span className='messagesUnreadCount'>{unreadNonFavoriteMessageCount}</span>}
                   </div>
                 </div>
-                <div className='collection__heading__handle' onClick={self.onClickChannelBrowser.bind(self, cabal.addr)}>
+                <div 
+                  className='collection__heading__handle' 
+                  onClick={this.onClickChannelBrowser.bind(this, cabal.addr)}
+                  title='Browse and join or create channels'
+                >
                   <img src='static/images/icon-newchannel.svg' />
                 </div>
               </div>
@@ -275,15 +294,23 @@ class SidebarScreen extends React.Component {
                 <div className='collection__heading__title__container'>
                   <span
                     className={`collection__toggle ${this.props.settings['sidebar-hide-peers'] ? 'collection__toggle__off' : 'collection__toggle__on'}`}
-                    onClick={self.onToggleCollection.bind(self, 'peers')}
+                    onClick={this.onToggleCollection.bind(this, 'peers')}
                   >▼
                   </span>
-                  <div className='collection__heading__title'>Peers - {onlineCount} online</div>
+                  <div
+                    className='collection__heading__title'
+                    onClick={this.onToggleCollection.bind(this, 'peers')}
+                  >
+                    Peers - {onlineCount} online
+                  </div>
                 </div>
                 <div className='collection__heading__handle' />
               </div>
               {!this.props.settings['sidebar-hide-peers'] && deduplicatedNicks.map((peer, index) => {
                 const keys = peer.users.map((u) => u.key).join(', ')
+                const isAdmin = peer.users.some((u) => u.isAdmin())
+                const isModerator = peer.users.some((u) => u.isModerator())
+                const isHidden = peer.users.some((u) => u.isHidden())
                 return (
                   <div
                     key={index}
@@ -298,9 +325,14 @@ class SidebarScreen extends React.Component {
                       {!peer.online &&
                         <img alt='Offline' src='static/images/icon-status-offline.svg' />}
                     </div>
-                    <div className={`collection__item__content ${peer.online ? 'active' : ''}`}>
-                      {peer.name || peer.key.substring(0, 6)}
-                      {peer.users.length > 1 && <span className='collection__item__count'>({peer.users.length})</span>}
+                    <div className={`collection__item__content ${(peer.online && !isHidden) ? 'active' : ''}`}>
+                      <span className='name'>
+                        {peer.name || peer.key.substring(0, 6)}
+                        {peer.users.length > 1 && <span className='collection__item__count'>({peer.users.length})</span>}
+                      </span>
+                      {!isAdmin && !isModerator && isHidden && <span className='sigil hidden'>HIDDEN</span>}
+                      {!isAdmin && isModerator && <span className='sigil moderator' title='Moderator'>MOD</span>}
+                      {isAdmin && <span className='sigil admin' title='Admin'>ADMIN</span>}
                     </div>
                     <div className='collection__item__handle' />
                   </div>