+++ app/reducer.js
@@ -11,6 +11,7 @@ const defaultState = {
   cabalSettings: {},
   cabalSettingsVisible: false,
   channelMembers: [],
+  channelPanelVisible: {},
   currentCabal: null,
   currentChannel: 'default',
   emojiPickerVisible: false,
@@ -76,12 +77,18 @@ const reducer = createReducer(defaultState, {
   UPDATE_WINDOW_BADGE: (state, { badgeCount }) => { state.badgeCount = badgeCount },
   SHOW_EMOJI_PICKER: (state) => { state.emojiPickerVisible = true },
   HIDE_EMOJI_PICKER: state => { state.emojiPickerVisible = false },
-  SHOW_PROFILE_PANEL: (state, { addr, user }) => {
+  SHOW_PROFILE_PANEL: (state, { addr, userKey }) => {
     state.profilePanelVisible[addr] = true
-    state.profilePanelUser[addr] = user
+    state.profilePanelUser[addr] = userKey
   },
   HIDE_PROFILE_PANEL: (state, { addr }) => {
     state.profilePanelVisible[addr] = false
+  },
+  SHOW_CHANNEL_PANEL: (state, { addr }) => {
+    state.channelPanelVisible[addr] = true
+  },
+  HIDE_CHANNEL_PANEL: (state, { addr }) => {
+    state.channelPanelVisible[addr] = false
   },
   UPDATE_SCREEN_VIEW_HISTORY: (state, { addr, channel }) => {
     state.screenViewHistory.push({ addr, channel })