+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [5.1.0] - 2020-06-06
+### Added
+- Added basic moderation features for hiding users and setting moderators and admins.
+- Added indicator for unread messages in collapsed channels list.
+- Optimized incoming message handling.
+- Added a dark mode theme.
+### Fixed
+- Fixed desktop notifications appearing from channels you have not joined. 
+
 ## [5.0.5] - 2020-05-19
 ### Added
 - Added "starred/favorite" channels list.